Firestore transactions

+++ server/js/cards.mjs
@@ -2,12 +2,13 @@ import Firestore from "@google-cloud/firestore"
 
 export default class {
   constructor(options) {
-    this.db = options.database
+    this.db = options.db
+    this.user = options.user
     this.search = options.search
   }
 
   get(id, success, failure) {
-    this.db.collection("Cards").doc(id).get().then(doc => {
+    this.user.collection("Cards").doc(id).get().then(doc => {
       if (doc.exists) {
         let card = doc.data()
         for (let field of [ "Date", "Hits", "Successes", "AvailableFrom", "LastHit", "Queue" ]) {
@@ -26,6 +27,7 @@ export default class {
     let today = new Date()
     let tomorrow  = new Date()
     let text = card["SearchPhrases"]
+    let doc = this.user.collection("Cards").doc()
     delete card["SearchPhrases"]
     tomorrow.setDate(today.getDate() + 1)
     card["Date"] = today
@@ -33,18 +35,13 @@ export default class {
     card["Revision"] = ""
     card["Successes"] = 0
     card["AvailableFrom"] = tomorrow
-    this.db.collection("Cards").add(card).then(doc => {
-      card["CardId"] = doc.id
-      return this.tags([], card["Tags"])
-    }).then(() => {
-      return new Promise((resolve, reject) => {
-        this.search.index(card["CardId"], card["Tags"], text, resolve, reject)
-      })
-    }).then(() => {
-      return this.db.update({
-        LastCard: card["Type"],
-        LastTags: card["Tags"]
-      })
+    this.db.runTransaction(transaction => {
+      let update = this.tags([], card["Tags"])
+      update["LastCard"] = card["Type"]
+      update["LastTags"] = card["Tags"]
+      transaction.set(doc, card)
+      transaction.update(this.user, update)
+      return this.search.index(doc.id, card["Tags"], text)
     }).then(() => {
       success("The card was successfully created")
     }).catch(error => {
@@ -53,34 +50,48 @@ export default class {
   }
 
   update(id, card, success, failure) {
-    this.get(id, (_, old) => {
-      let text = card["SearchPhrases"]
-      delete card["SearchPhrases"]
-      card["ScheduledFor"] = card["ScheduledFor"] || Firestore.FieldValue.delete()
-      this.db.collection("Cards").doc(id).set(card, { merge: true }).then(() => {
-        return this.tags(old["Tags"], card["Tags"])
-      }).then(() => {
-        return new Promise((resolve, reject) => {
-          this.search.index(id, card["Tags"], text, resolve, reject)
-        })
-      }).then(() => {
-        success("The card was successfully updated")
-      }).catch(error => {
-        failure("Failed to updated the card", { id, error })
+    let ref = this.user.collection("Cards").doc(id)
+    this.db.runTransaction(transaction => {
+      return transaction.get(ref).then(doc => {
+        if (!doc.exists) {
+          throw new Error("Card " + id + " does not exist")
+        }
+        let text = card["SearchPhrases"]
+        delete card["SearchPhrases"]
+        card["ScheduledFor"] = card["ScheduledFor"] || Firestore.FieldValue.delete()
+        transaction.set(ref, card, { merge: true })
+        let tags = this.tags(doc.data()["Tags"], card["Tags"])
+        if (Object.keys(tags).length > 0) {
+          transaction.update(this.user, tags)
+        }
+        return this.search.index(id, card["Tags"], text)
       })
-    }, failure)
+    }).then(() => {
+      success("The card was successfully updated")
+    }).catch(error => {
+      failure("Failed to updated the card", { id, error })
+    })
   }
 
   remove(id, success, failure) {
-    this.get(id, (_, card) => {
-      this.db.collection("Cards").doc(id).delete().then(() => {
-        return this.tags(card["Tags"], [])
-      }).then(() => {
-        success("The card was successfully deleted")
-      }).catch(error => {
-        failure("Failed to delete the card", { id, error })
+    let ref = this.user.collection("Cards").doc(id)
+    this.db.runTransaction(transaction => {
+      return transaction.get(ref).then(doc => {
+        if (!doc.exists) {
+          throw new Error("Card " + id + " does not exist")
+        }
+        transaction.delete(ref)
+        let tags = this.tags(doc.data()["Tags"], [])
+        if (Object.keys(tags).length > 0) {
+          transaction.update(this.user, tags)
+        }
+        return this.search.remove(id)
       })
-    }, failure)
+    }).then(() => {
+      success("The card was successfully deleted")
+    }).catch(error => {
+      failure("Failed to delete the card", { id, error })
+    })
   }
 
   tags(remove, add) {
@@ -98,8 +109,6 @@ export default class {
         tags[key] = Firestore.FieldValue.increment(1)
       }
     }
-    if (Object.keys(tags).length > 0) {
-      return this.db.update(tags)
-    }
+    return tags
   }
 }
